+++ Readme.md
@@ -29,6 +29,13 @@ Then add the following line to $HOME/Secrets/dbpass.txt again putting the passwo
 emag    jdbc:postgresql://127.0.0.1:5432/emag       emag        password
 ```
 
+Make sure each field is separated by a single TAB character.
+
+To quickly drop all tables for testing, execute this:
+
+```
+psql -d emag -c "SELECT 'DROP TABLE IF EXISTS ' || tablename || ' CASCADE;' FROM pg_tables WHERE schemaname = 'public';" | psql -d emag
+```
 
 Stuff related to scraping app
 =
